feat: Enhance data generation capabilities and update dependencies
- Moved `generateFakeData.js` file from `src/lib` to a new location  for better management.
- Added a new `generateData.js` script for manual data generation via npm command.
- Updated `package.json` to include `@trivago/prettier-plugin-sort-imports` and other dependencies.
- Added mock API functions for fetching various types of data, including weather and booking statistics.
