+++ package.json
@@ -1,60 +1,63 @@
 {
-  "name": "hotel-booking-nextjs",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@hookform/resolvers": "^3.9.0",
-    "@radix-ui/react-accordion": "^1.2.2",
-    "@radix-ui/react-avatar": "^1.1.2",
-    "@radix-ui/react-dialog": "^1.1.2",
-    "@radix-ui/react-dropdown-menu": "^2.1.1",
-    "@radix-ui/react-label": "^2.1.0",
-    "@radix-ui/react-popover": "^1.1.1",
-    "@radix-ui/react-scroll-area": "^1.2.2",
-    "@radix-ui/react-select": "^2.1.1",
-    "@radix-ui/react-separator": "^1.1.0",
-    "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-switch": "^1.1.2",
-    "@radix-ui/react-tabs": "^1.1.2",
-    "@radix-ui/react-tooltip": "^1.1.4",
-    "axios": "^1.7.9",
-    "bcryptjs": "^2.4.3",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
-    "cmdk": "^1.0.4",
-    "date-fns": "^3.6.0",
-    "dotenv": "^16.4.7",
-    "framer-motion": "^11.5.4",
-    "jose": "^5.9.3",
-    "js-cookie": "^3.0.5",
-    "jsonwebtoken": "^9.0.2",
-    "lucide-react": "^0.428.0",
-    "mongodb": "^6.8.0",
-    "mongoose": "^8.6.0",
-    "next": "^14.2.7",
-    "react": "^18",
-    "react-day-picker": "^8.10.1",
-    "react-dom": "^18",
-    "react-hook-form": "^7.53.0",
-    "recharts": "^2.15.0",
-    "tailwind-merge": "^2.5.2",
-    "tailwindcss-animate": "^1.0.7",
-    "zod": "^3.23.8"
-  },
-  "devDependencies": {
-    "@faker-js/faker": "^9.4.0",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.5",
-    "postcss": "^8",
-    "prettier": "3.4.2",
-    "prettier-plugin-tailwindcss": "^0.6.9",
-    "tailwindcss": "^3.4.1"
-  }
+	"name": "hotel-booking-nextjs",
+	"version": "0.1.0",
+	"private": true,
+	"type": "module",
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
+		"lint": "next lint",
+		"prettier": "npx prettier ./src --write",
+    "generate": "node src/generateData.js"
+	},
+	"dependencies": {
+		"@hookform/resolvers": "^3.9.0",
+		"@radix-ui/react-accordion": "^1.2.2",
+		"@radix-ui/react-avatar": "^1.1.2",
+		"@radix-ui/react-dialog": "^1.1.2",
+		"@radix-ui/react-dropdown-menu": "^2.1.1",
+		"@radix-ui/react-label": "^2.1.0",
+		"@radix-ui/react-popover": "^1.1.1",
+		"@radix-ui/react-scroll-area": "^1.2.2",
+		"@radix-ui/react-select": "^2.1.1",
+		"@radix-ui/react-separator": "^1.1.0",
+		"@radix-ui/react-slot": "^1.1.0",
+		"@radix-ui/react-switch": "^1.1.2",
+		"@radix-ui/react-tabs": "^1.1.2",
+		"@radix-ui/react-tooltip": "^1.1.4",
+		"axios": "^1.7.9",
+		"bcryptjs": "^2.4.3",
+		"class-variance-authority": "^0.7.1",
+		"clsx": "^2.1.1",
+		"cmdk": "^1.0.4",
+		"date-fns": "^3.6.0",
+		"dotenv": "^16.4.7",
+		"framer-motion": "^11.5.4",
+		"jose": "^5.9.3",
+		"js-cookie": "^3.0.5",
+		"jsonwebtoken": "^9.0.2",
+		"lucide-react": "^0.428.0",
+		"mongodb": "^6.8.0",
+		"mongoose": "^8.6.0",
+		"next": "^14.2.7",
+		"react": "^18",
+		"react-day-picker": "^8.10.1",
+		"react-dom": "^18",
+		"react-hook-form": "^7.53.0",
+		"recharts": "^2.15.0",
+		"tailwind-merge": "^2.5.2",
+		"tailwindcss-animate": "^1.0.7",
+		"zod": "^3.23.8"
+	},
+	"devDependencies": {
+		"@faker-js/faker": "^9.4.0",
+		"@trivago/prettier-plugin-sort-imports": "^5.2.1",
+		"eslint": "^8",
+		"eslint-config-next": "14.2.5",
+		"postcss": "^8",
+		"prettier": "3.4.2",
+		"prettier-plugin-tailwindcss": "^0.6.9",
+		"tailwindcss": "^3.4.1"
+	}
 }
